==> 02_asets_regressionAnalysis.Rmd <==
---
title: "Step 2: Regression analysis with brms"
subtitle: "Longitudinal association of homonegative school climate with body dysmorphic disorder among cisgender sexual minority adolescents: Testing mediational effects of proximal stressors"
author: "F. Hunter McGuire, MPH"
date: "`r format(Sys.Date(), format='%B %d, %Y')`"
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: false
    theme: readable
    highlight: zenburn
urlcolor: blue
editor_options: 
  markdown: 
    wrap: 72
  chunk_output_type: console
---

# Load data and packages

This first code chunk loads R packages, sets the working directory to
the source file location, and loads the long-format, imputed data frame
including cisgender sexual minority participants.

```{r setup, message=FALSE, warning=FALSE}

knitr::opts_chunk$set(warning=FALSE, message=FALSE)

# load packages
library(tidyr) # data management tasks
library(dplyr) # data management tasks
library(mice) # multiple imputation with chained equations
library(brms) # Bayesian regression modeling with Stan

## set working directory to source file location
### get full file path
RMDpath <- rstudioapi::getSourceEditorContext()$path
### remove file name from file path
RMDpath <- gsub('/02_asets_regressionAnalysis.Rmd','', RMDpath)
### set the working directory
setwd(RMDpath)

# load dataset
imputed_long_cisgender <- readRDS(file="imputed_long_cisgender.RDS")

```

# Main analysis

## Define regression formulas

The following code defines the regression formulas used for the primary
analysis presented in Table 3 of the journal article. For clarity these
are referred in the journal article as:

-   Model 1 = "totalEffectModel"

-   Model 2 = "internalizedHomonegativityModel"

-   Model 3 = "negativeExpectanciesModel"

-   Model 4 = "indirectOverallModel"

In each model, I have specified random intercepts and random slopes for
all variables except grand mean-centered age. The random effects allow
for participant-specific variation around the fixed effects estimates. I
also use Poisson outcome distributions with log link functions to
estimate risk ratios from exponentiated fixed effects coefficients.

For example, for individual participant *i* at time *t*, the first model
(Model 1 or "totalEffectModel") is specified as:

**Level 1:**

$y_{ti} = \beta_{0i} + \beta_{1i}Time + \beta_{2i}(HSC_{ti} - \overline{HSC_i}) + \beta_{3i}(Age_{ti} - \overline{Age})$

**Level 2:**

$\beta_{0i} = \gamma_{00} + \gamma_{01}\overline{HSC_i} + u_{0i}$

$\beta_{1i} = \gamma_{10} + u_{1i}$

$\beta_{2i} = \gamma_{20} + u_{2i}$

$\beta_{3i} = \gamma_{30}$

```{r mainAnalysis_defineModels}

# Total effect of homonegative school climate
totalEffectModel <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct + 
       grandMeanCenter_age + (1 + time + personMeanCenter_hscpct | pid), 
     family=poisson(link="log"))

# Partial indirect effect through internalized homonegativity
internalizedHomonegativityModel <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct +
       personMeanCenter_ihpct + personMean_ihpct + grandMeanCenter_age +
       (1 + time + personMeanCenter_hscpct + personMeanCenter_ihpct | pid), 
     family=poisson(link="log"))

# Partial indirect effect through negative expectancies
negativeExpectanciesModel <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct +
       personMeanCenter_nepct + personMean_nepct + grandMeanCenter_age +
       (1 + time + personMeanCenter_hscpct + personMeanCenter_nepct | pid), 
     family=poisson(link="log"))

# Overall indirect effect via internalized homoneg. & neg. expectancies
indirectOverallModel <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct +
       personMeanCenter_ihpct + personMean_ihpct + personMeanCenter_nepct +
       personMean_nepct + grandMeanCenter_age + 
       (1 + time + personMeanCenter_hscpct +
          personMeanCenter_ihpct + personMeanCenter_nepct | pid),
     family=poisson(link="log"))

```

## Fit regression models with `brms`

The following code uses the imputed data frame and the defined
regression models as inputs. To estimate the model parameters, I use
Markov chain Monte Carlo (MCMC) sampling using the `brm_multiple`
function from `brms`. This function runs analysis separately on each
imputation and pools results at the end. A great tutorial on this
function can be found
[here](https://paul-buerkner.github.io/brms/articles/brms_missings.html).

For each model, I have specified four chains, 2000 warmup iterations,
and 4000 total iterations (i.e., 2000 post-warmup samples).

```{r mainAnalysis_fit}

# run models in parallel
future::plan("multisession", workers=5) 
fit_totalEffect <-
  brm_multiple(totalEffectModel, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/main/fit_totalEffect.rds")

# run models in parallel
future::plan("multisession", workers=5) 
fit_internalizedHomonegativity <-
  brm_multiple(internalizedHomonegativityModel, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/main/fit_internalizedHomonegativity.rds")

# run models in parallel
future::plan("multisession", workers=5) 
fit_negativeExpectancies <-
  brm_multiple(negativeExpectanciesModel, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/main/fit_negativeExpectancies.rds")

# run models in parallel
future::plan("multisession", workers=5) 
fit_indirectOverall <-
  brm_multiple(indirectOverallModel, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/main/fit_indirectOverall.rds")

```

# Sensitivity 1: Add race/ethnicity and gender as level 2 predictor variables

Race/ethnicity and gender are time-invariant, so they are included as
level 2 predictors of the model intercept.

## Define regression model formulas

```{r sensitivity1_raceGender_defineModels}
# Total effect of homonegative school climate
totalEffectModel_raceGender <- 
  bf(bddbin2 ~ 1 +  time + personMeanCenter_hscpct + personMean_hscpct +
       grandMeanCenter_age + cisgenderBoy +
       asian + black + hispanic + otherMultiracial +
       (1 + time + personMeanCenter_hscpct | pid), 
     family=poisson(link="log"))
# Partial indirect effect through internalized homonegativity
internalizedHomonegativityModel_raceGender <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct +
       personMeanCenter_ihpct + personMean_ihpct + grandMeanCenter_age +
       cisgenderBoy + asian + black + hispanic + otherMultiracial +
       (1 + time + personMeanCenter_hscpct + personMeanCenter_ihpct | pid),
     family=poisson(link="log"))
# Partial indirect effect through negative expectancies
negativeExpectanciesModel_raceGender <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct +
       personMeanCenter_nepct + personMean_nepct + grandMeanCenter_age +
       cisgenderBoy + asian + black + hispanic + otherMultiracial +
       (1 + time + personMeanCenter_hscpct + personMeanCenter_nepct | pid),
     family=poisson(link="log"))
# Overall indirect effect via internalized homoneg. & neg. expectancies
indirectOverallModel_raceGender <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscpct + personMean_hscpct + 
       personMeanCenter_ihpct + personMean_ihpct + 
       personMeanCenter_nepct + personMean_nepct + grandMeanCenter_age +
       cisgenderBoy + asian + black + hispanic + otherMultiracial +
       (1 + time + personMeanCenter_hscpct + personMeanCenter_ihpct + 
          personMeanCenter_nepct | pid),
     family=poisson(link="log"))

```

## Fit regression models with `brms`

```{r sensitivity1_raceGender_fit}

# Model 1
future::plan("multisession", workers=5) 
fit_totalEffect_raceGender <-
  brm_multiple(totalEffectModel_raceGender, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity1_raceGender/fit_totalEffect_raceGender.rds")

# Model 2
future::plan("multisession", workers=5) 
fit_internalizedHomonegativity_raceGender <-
  brm_multiple(internalizedHomonegativityModel_raceGender, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity1_raceGender/fit_internalizedHomonegativity_raceGender.rds")

# Model 3
future::plan("multisession", workers=5) 
fit_negativeExpectancies_raceGender <-
  brm_multiple(negativeExpectanciesModel_raceGender, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity1_raceGender/fit_negativeExpectancies_raceGender.rds")

# Model 4
future::plan("multisession", workers=5) 
fit_indirectOverall_raceGender <-
  brm_multiple(indirectOverallModel_raceGender, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity1_raceGender/fit_indirectOverall_raceGender.rds")

```

# Sensitivity 2: Binary minority stress variables

## Define regression model formulas

```{r sensitivity2_binary_defineModels}

# Total effect of homonegative school climate
totalEffectModel_binary <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscbinary + personMean_hscbinary +
       grandMeanCenter_age +
       (1 + time + personMeanCenter_hscbinary | pid), 
     family=poisson(link="log"))
# Partial indirect effect through internalized homonegativity
internalizedHomonegativityModel_binary <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscbinary + personMean_hscbinary + 
       personMeanCenter_ihbinary + personMean_ihbinary +
       grandMeanCenter_age +
       (1 + time + personMeanCenter_hscbinary + 
          personMeanCenter_ihbinary | pid),
     family=poisson(link="log"))
# Partial indirect effect through negative expectancies
negativeExpectanciesModel_binary <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscbinary + personMean_hscbinary + 
       personMeanCenter_nebinary + personMean_nebinary +
       grandMeanCenter_age +
       (1 + time + personMeanCenter_hscbinary + 
          personMeanCenter_nebinary | pid),
     family=poisson(link="log"))
# Overall indirect effect via internalized homoneg. & neg. expectancies
indirectOverallModel_binary <- 
  bf(bddbin2 ~ 1 + time + personMeanCenter_hscbinary + personMean_hscbinary +
       personMeanCenter_ihbinary + personMean_ihbinary +
       personMeanCenter_nebinary + personMean_nebinary +
       grandMeanCenter_age +
       (1 + time + personMeanCenter_hscbinary + 
          personMeanCenter_ihbinary + personMeanCenter_nebinary | pid),
     family=poisson(link="log"))

```

## Fit regression models with `brms`

```{r sensitivity2_binary_fit}

# Model 1
future::plan("multisession", workers=5) 
fit_totalEffect_binary <-
  brm_multiple(totalEffectModel_binary, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity2_binary/fit_totalEffect_binary.rds")

# Model 2
future::plan("multisession", workers=5) 
fit_internalizedHomonegativity_binary <-
  brm_multiple(internalizedHomonegativityModel_binary, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity2_binary/fit_internalizedHomonegativity_binary.rds")

# Model 3
future::plan("multisession", workers=5) 
fit_negativeExpectancies_binary <-
  brm_multiple(negativeExpectanciesModel_binary, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity2_binary/fit_negativeExpectancies_binary.rds")

# Model 4
future::plan("multisession", workers=5) 
fit_indirectOverall_binary <-
  brm_multiple(indirectOverallModel_binary, 
               data = imputed_long_cisgender,
               seed = 828,
               chains = 4,
               warmup = 2000,
               iter = 4000,
               thin = 1,
               cores = 8,
               control = list(adapt_delta = 0.95),
               file = "fits/sensitivity2_binary/fit_indirectOverall_binary.rds")

```
