==> 04_asets_sensitivity.Rmd <==
---
title: "Step 4: Sensitivity analysis"
subtitle: "Longitudinal association of homonegative school climate with body dysmorphic disorder among cisgender sexual minority adolescents: Testing mediational effects of proximal stressors"
author: "F. Hunter McGuire, MPH"
date: "`r format(Sys.Date(), format='%B %d, %Y')`"
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: false
    theme: readable
    highlight: zenburn
urlcolor: blue
editor_options: 
  markdown: 
    wrap: 72
  chunk_output_type: console
---

# Load model fits and packages

This first code chunk loads R packages, sets the working directory to
the source file location, and loads the sensitivity analysis `brms`
model fits.

```{r setup, message=FALSE, warning=FALSE}

knitr::opts_chunk$set(warning=FALSE, message=FALSE)

# load packages
library(tidyr) # data management tasks
library(dplyr) # data management tasks
library(brms) # Bayesian regression modeling with Stan
library(mice) # multiple imputation

## set working directory to source file location
### get full file path
RMDpath <- rstudioapi::getSourceEditorContext()$path
### remove file name from file path
RMDpath <- gsub('/04_asets_sensitivity.Rmd','', RMDpath)
### set the working directory
setwd(RMDpath)

# load model fits for sensitivity 1 (add race & gender fixed effects)
fit_totalEffect_raceGender <- 
  readRDS(file="fits/sensitivity1_raceGender/fit_totalEffect_raceGender.rds")
fit_internalizedHomonegativity_raceGender <- 
  readRDS(file="fits/sensitivity1_raceGender/fit_internalizedHomonegativity_raceGender.rds")
fit_negativeExpectancies_raceGender <- 
  readRDS(file="fits/sensitivity1_raceGender/fit_negativeExpectancies_raceGender.rds")
fit_indirectOverall_raceGender <- 
  readRDS(file="fits/sensitivity1_raceGender/fit_indirectOverall_raceGender.rds")

# load model fits for sensitivity 2 (binary predictors)
fit_totalEffect_binary <- 
  readRDS(file="fits/sensitivity2_binary/fit_totalEffect_binary.rds")
fit_internalizedHomonegativity_binary <- 
  readRDS(file="fits/sensitivity2_binary/fit_internalizedHomonegativity_binary.rds")
fit_negativeExpectancies_binary <- 
  readRDS(file="fits/sensitivity2_binary/fit_negativeExpectancies_binary.rds")
fit_indirectOverall_binary <- 
  readRDS(file="fits/sensitivity2_binary/fit_indirectOverall_binary.rds")

```

# Sensitivity analysis 1 (S1): Race and gender-adjusted models

This code extracts information from the Sensitivity Analysis 1 `brms`
model fits and organizes this information into neatly formatted excel
file.

## Calcluate fixed effects coefficients (Risk ratios)

```{r fixedEffects_s1}
fixedEffects_s1 <- as.data.frame(matrix(nrow=16, ncol=3))
colnames(fixedEffects_s1) <- c("RR", "LB", "UB")
rownames(fixedEffects_s1) <-  c("Model 1: pmc HSC",
                                "Model 1: pm HSC",
                                "Model 2: pmc HSC", 
                                "Model 2: pmc IH",
                                "Model 2: pm HSC",
                                "Model 2: pm IH",
                                "Model 3: pmc HSC",
                                "Model 3: pmc NE",
                                "Model 3: pm HSC",
                                "Model 3: pm NE",
                                "Model 4: pmc HSC",
                                "Model 4: pmc IH",
                                "Model 4: pmc NE",
                                "Model 4: pm HSC",
                                "Model 4: pm IH",
                                "Model 4: pm NE")
# Model 1
fixedEffects_s1[1,] <- exp(fixef(fit_totalEffect_raceGender)[3,c(1,3,4)])
fixedEffects_s1[2,] <- exp(fixef(fit_totalEffect_raceGender)[4,c(1,3,4)])
# Model 2
fixedEffects_s1[3,] <- 
  exp(fixef(fit_internalizedHomonegativity_raceGender)[3,c(1,3,4)])
fixedEffects_s1[4,] <- 
  exp(fixef(fit_internalizedHomonegativity_raceGender)[5,c(1,3,4)])
fixedEffects_s1[5,] <- 
  exp(fixef(fit_internalizedHomonegativity_raceGender)[4,c(1,3,4)])
fixedEffects_s1[6,] <- 
  exp(fixef(fit_internalizedHomonegativity_raceGender)[6,c(1,3,4)])
# Model 3
fixedEffects_s1[7,] <- 
  exp(fixef(fit_negativeExpectancies_raceGender)[3,c(1,3,4)])
fixedEffects_s1[8,] <- 
  exp(fixef(fit_negativeExpectancies_raceGender)[5,c(1,3,4)])
fixedEffects_s1[9,] <- 
  exp(fixef(fit_negativeExpectancies_raceGender)[4,c(1,3,4)])
fixedEffects_s1[10,] <- 
  exp(fixef(fit_negativeExpectancies_raceGender)[6,c(1,3,4)])
# Model 4
fixedEffects_s1[11,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[3,c(1,3,4)])
fixedEffects_s1[12,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[5,c(1,3,4)])
fixedEffects_s1[13,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[7,c(1,3,4)])
fixedEffects_s1[14,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[4,c(1,3,4)])
fixedEffects_s1[15,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[6,c(1,3,4)])
fixedEffects_s1[16,] <- 
  exp(fixef(fit_indirectOverall_raceGender)[8,c(1,3,4)])
# combine means and 95% CI into one column
fixedEffects_s1 <- fixedEffects_s1 %>% 
  mutate(combined = 
           paste(
             format(round(RR,2),nsmall=2), 
             " (", 
             format(round(LB,2),nsmall=2),
             ", ", 
             format(round(UB,2),nsmall=2),
             ")", 
             sep=""))

```

## Calculate Level 2 variance estimates

```{r betweenPersonVariance_s1}

# define parameter names
HSCvar <- "sd_pid__personMeanCenter_hscpct"
IHvar <- "sd_pid__personMeanCenter_ihpct"
NEvar <- "sd_pid__personMeanCenter_nepct"

# extract level 2 variance for each parameter
# squared (^2) because the default is standard deviation
model1_HSCvar_s1 <- 
  posterior_samples(fit_totalEffect_raceGender, 
                     pars=c(HSCvar))[[1]]^2
model2_HSCvar_s1 <- 
  posterior_samples(fit_internalizedHomonegativity_raceGender, 
                     pars=c(HSCvar))[[1]]^2
model2_IHvar_s1 <- 
  posterior_samples(fit_internalizedHomonegativity_raceGender, 
                     pars=c(IHvar))[[1]]^2
model3_HSCvar_s1 <- 
  posterior_samples(fit_negativeExpectancies_raceGender, 
                     pars=c(HSCvar))[[1]]^2
model3_NEvar_s1 <- 
  posterior_samples(fit_negativeExpectancies_raceGender, 
                     pars=c(NEvar))[[1]]^2
model4_HSCvar_s1 <- 
  posterior_samples(fit_indirectOverall_raceGender, 
                     pars=c(HSCvar))[[1]]^2
model4_IHvar_s1 <- 
  posterior_samples(fit_indirectOverall_raceGender, 
                     pars=c(IHvar))[[1]]^2
model4_NEvar_s1 <- 
  posterior_samples(fit_indirectOverall_raceGender, 
                     pars=c(NEvar))[[1]]^2



# Create a matrix called "variance" to hold Level 2 variance estimates
variance_s1 <- matrix(nrow=8, ncol=1)
colnames(variance_s1) <- c("variance")
rownames(variance_s1) <-  c("Model 1: HSC",
                            "Model 2: HSC", 
                            "Model 2: IH",
                            "Modle 3: HSC",
                            "Model 3: NE",
                            "Model 4: HSC",
                            "Model 4: IH",
                            "Model 4: NE")
# assign variance values to each cell
variance_s1[1] <- round(mean(model1_HSCvar_s1),4)
variance_s1[2] <- round(mean(model2_HSCvar_s1),4)
variance_s1[3] <- round(mean(model2_IHvar_s1),4)
variance_s1[4] <- round(mean(model3_HSCvar_s1),4)
variance_s1[5] <- round(mean(model3_NEvar_s1),4)
variance_s1[6] <- round(mean(model4_HSCvar_s1),4)
variance_s1[7] <- round(mean(model4_IHvar_s1),4)
variance_s1[8] <- round(mean(model4_NEvar_s1),4)

```

## Calcluate indirect effect estimates and percentage mediated

```{r extractPosteriorDistributions_s1}

# Name of the primary parameter of interest
# Person mean-centered homonegative school climate
HSC <- "b_personMeanCenter_hscpct"

# Extract estimates for this parameter from each model fit into numeric vector
model1_beta_samples_s1 <- 
  (posterior_samples(fit_totalEffect_raceGender, 
                    pars=c(HSC)))[[1]]
model2_beta_samples_s1 <- 
  posterior_samples(fit_internalizedHomonegativity_raceGender,
                    pars=c(HSC))[[1]]
model3_beta_samples_s1 <- 
  posterior_samples(fit_negativeExpectancies_raceGender, 
                    pars=c(HSC))[[1]]
model4_beta_samples_s1 <- 
  posterior_samples(fit_indirectOverall_raceGender, 
                    pars=c(HSC))[[1]]

# Use the difference method to calculate indirect effects
# subtract Model 2-4 beta post. distributions from Model 1 post. distributions
model2_difference_s1 <- model1_beta_samples_s1 - model2_beta_samples_s1
model3_difference_s1 <- model1_beta_samples_s1 - model3_beta_samples_s1
model4_difference_s1 <- model1_beta_samples_s1 - model4_beta_samples_s1

# calculate percentage mediated for each model (relative to Model 1)
model2_percent_s1 <- model2_difference_s1/model1_beta_samples_s1
model3_percent_s1 <- model3_difference_s1/model1_beta_samples_s1
model4_percent_s1 <- model4_difference_s1/model1_beta_samples_s1


# Create a matrix to hold indirect effect estimates
indirect_s1 <- as.data.frame(matrix(nrow=3, ncol=3))
colnames(indirect_s1) <- c("mean", "LB", "UB")
rownames(indirect_s1) <-  c("Model 2",
                            "Model 3",
                            "Model 4")
# Calculate indirect effect estimates for Models 2, 3, and 4
# Model 2
indirect_s1[1,1] <- exp(mean(model2_difference_s1))
indirect_s1[1,2:3] <- exp(quantile(model2_difference_s1, 
                                      probs=c(0.025,0.975)))
# Model 3
indirect_s1[2,1] <- exp(mean(model3_difference_s1))
indirect_s1[2,2:3] <- exp(quantile(model3_difference_s1, 
                                      probs=c(0.025,0.975)))
# Model 4
indirect_s1[3,1] <- exp(mean(model4_difference_s1))
indirect_s1[3,2:3] <- exp(quantile(model4_difference_s1, 
                                      probs=c(0.025,0.975)))
# combine means and 95% CI into one column
indirect_s1 <- indirect_s1 %>% 
  mutate(combined = 
           paste(
             format(round(mean,2),nsmall=2), 
             " (", 
             format(round(LB,2),nsmall=2),
             ", ", 
             format(round(UB,2),nsmall=2),
             ")", 
             sep=""))



# Create a matrix to hold percentage mediated estimates
percent_s1 <- as.data.frame(matrix(nrow=3, ncol=3))
colnames(percent_s1) <- c("mean", "LB", "UB")
rownames(percent_s1) <-  c("Model 2",
                           "Model 3",
                           "Model 4")
# Calculate percentage mediated for Models 2, 3, and 4
# Model 2
percent_s1[1,1] <- mean(model2_percent_s1)*100
percent_s1[1,2:3] <- quantile(model2_percent_s1, probs=c(0.025,0.975))*100
# Model 3
percent_s1[2,1] <- mean(model3_percent_s1)*100
percent_s1[2,2:3] <- quantile(model3_percent_s1, probs=c(0.025,0.975))*100
# Model 4 
percent_s1[3,1] <- mean(model4_percent_s1)*100
percent_s1[3,2:3] <- quantile(model4_percent_s1, probs=c(0.025,0.975))*100
# combine means and 95% CI into one column
percent_s1 <- percent_s1 %>% 
  mutate(combined = 
           paste(
             format(round(mean,1),nsmall=1), 
             "% (", 
             format(round(LB,1),nsmall=1),
             ", ", 
             format(round(UB,1),nsmall=1),
             ")", 
             sep=""))

```

## Create and Export Table S1

```{r tableS1}

# parameter names
tableS1_params <-   c("Model 1",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "",
                      "Model 2",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "PMC Internalized homonegativity, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)",
                      "",
                      "Model 3",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "PMC Negative expectancies, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)",
                      "",
                      "Model 4",
                      "Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "Internalized homonegativity, RR (95% CI)",
                      "    Random slopes variance",
                      "Negative expectancies, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)")
# model estimates
tableS1_estimates <-  c("",
                        fixedEffects_s1[1,4],
                        variance_s1[1],
                        "",
                        "",
                        fixedEffects_s1[3,4],
                        variance_s1[2],
                        fixedEffects_s1[4,4],
                        variance_s1[3],
                        percent_s1[1,4],
                        "",
                        "",
                        fixedEffects_s1[7,4],
                        variance_s1[4],
                        fixedEffects_s1[8,4],
                        variance_s1[5],
                        percent_s1[2,4],
                        "",
                        "",
                        fixedEffects_s1[11,4],
                        variance_s1[6],
                        fixedEffects_s1[12,4],
                        variance_s1[7],
                        fixedEffects_s1[13,4],
                        variance_s1[8],
                        percent_s1[3,4])
# combine columns into one table
tableS1 <- as.data.frame(cbind(tableS1_params, tableS1_estimates))
# change column names
colnames(tableS1) <- c("Parameters", "Estimates")

# load library
library(openxlsx)

# save xlsx file
openxlsx::write.xlsx(tableS1, file="supplemental_files/tableS1.xlsx")
# reload into R in xlsx format
wb <- openxlsx::loadWorkbook(file="supplemental_files/tableS1.xlsx")
# set column width for PARAMETER to 40
openxlsx::setColWidths(wb, 
                       sheet = 1,
                       cols = 1, 
                       widths=50)
# set column width for model estimates to 20
openxlsx::setColWidths(wb, 
                       sheet = 1,
                       cols = 2, 
                       widths=30)
# change font size to 10 with Arial style
modifyBaseFont(wb, 
               fontSize = 10, 
               fontColour = "black", 
               fontName = "Times New Roman")

# bold text
boldStyle <- createStyle(textDecoration = "Bold")
# center-aligned text
centerStyle <- createStyle(halign = "CENTER")
# bold and center-aligned
boldCenterStyle <- createStyle(textDecoration = "Bold",
                               halign = "CENTER")

# apply styles
addStyle(wb,
         rows = c(1,2,6,13,20),
         cols = 1,
         sheet = 1,
         style = boldStyle)
addStyle(wb,
         cols = 2,
         rows = 2:27,
         sheet = 1,
         style = centerStyle,
         gridExpand = TRUE)
addStyle(wb,
         rows = c(1,3,7,9,11,14,16,21,23,27),
         cols = 2,
         sheet = 1,
         style = boldCenterStyle,
         gridExpand = TRUE)

# save the final excel workbook
openxlsx::saveWorkbook(wb, 
                         file="supplemental_files/tableS1.xlsx", 
                       overwrite = TRUE)

```

# Sensitivity analysis 2 (S2): Binary minority stress predictors

This code extracts information from the Sensitivity Analysis 2 `brms`
model fits and organizes this information into neatly formatted excel
file.

## Calcluate fixed effects coefficients (Risk ratios)

```{r fixedEffects_s2}
fixedEffects_s2 <- as.data.frame(matrix(nrow=16, ncol=3))
colnames(fixedEffects_s2) <- c("RR", "LB", "UB")
rownames(fixedEffects_s2) <-  c("Model 1: pmc HSC",
                                "Model 1: pm HSC",
                                "Model 2: pmc HSC", 
                                "Model 2: pmc IH",
                                "Model 2: pm HSC",
                                "Model 2: pm IH",
                                "Model 3: pmc HSC",
                                "Model 3: pmc NE",
                                "Model 3: pm HSC",
                                "Model 3: pm NE",
                                "Model 4: pmc HSC",
                                "Model 4: pmc IH",
                                "Model 4: pmc NE",
                                "Model 4: pm HSC",
                                "Model 4: pm IH",
                                "Model 4: pm NE")
# Model 1
fixedEffects_s2[1,] <- exp(fixef(fit_totalEffect_binary)[3,c(1,3,4)])
fixedEffects_s2[2,] <- exp(fixef(fit_totalEffect_binary)[4,c(1,3,4)])
# Model 2
fixedEffects_s2[3,] <- 
  exp(fixef(fit_internalizedHomonegativity_binary)[3,c(1,3,4)])
fixedEffects_s2[4,] <- 
  exp(fixef(fit_internalizedHomonegativity_binary)[5,c(1,3,4)])
fixedEffects_s2[5,] <- 
  exp(fixef(fit_internalizedHomonegativity_binary)[4,c(1,3,4)])
fixedEffects_s2[6,] <- 
  exp(fixef(fit_internalizedHomonegativity_binary)[6,c(1,3,4)])
# Model 3
fixedEffects_s2[7,] <- 
  exp(fixef(fit_negativeExpectancies_binary)[3,c(1,3,4)])
fixedEffects_s2[8,] <- 
  exp(fixef(fit_negativeExpectancies_binary)[5,c(1,3,4)])
fixedEffects_s2[9,] <- 
  exp(fixef(fit_negativeExpectancies_binary)[4,c(1,3,4)])
fixedEffects_s2[10,] <- 
  exp(fixef(fit_negativeExpectancies_binary)[6,c(1,3,4)])
# Model 4
fixedEffects_s2[11,] <- 
  exp(fixef(fit_indirectOverall_binary)[3,c(1,3,4)])
fixedEffects_s2[12,] <- 
  exp(fixef(fit_indirectOverall_binary)[5,c(1,3,4)])
fixedEffects_s2[13,] <- 
  exp(fixef(fit_indirectOverall_binary)[7,c(1,3,4)])
fixedEffects_s2[14,] <- 
  exp(fixef(fit_indirectOverall_binary)[4,c(1,3,4)])
fixedEffects_s2[15,] <- 
  exp(fixef(fit_indirectOverall_binary)[6,c(1,3,4)])
fixedEffects_s2[16,] <- 
  exp(fixef(fit_indirectOverall_binary)[8,c(1,3,4)])
# combine means and 95% CI into one column
fixedEffects_s2 <- fixedEffects_s2 %>% 
  mutate(combined = 
           paste(
             format(round(RR,2),nsmall=2), 
             " (", 
             format(round(LB,2),nsmall=2),
             ", ", 
             format(round(UB,2),nsmall=2),
             ")", 
             sep=""))

```

## Calculate Level 2 variance estimates

```{r betweenPersonVariance_s2}

# define parameter names
HSCvar <- "sd_pid__personMeanCenter_hscbinary"
IHvar <- "sd_pid__personMeanCenter_ihbinary"
NEvar <- "sd_pid__personMeanCenter_nebinary"

# extract level 2 variance for each parameter
# squared (^2) because the default is standard deviation
model1_HSCvar_s2 <- 
  posterior_samples(fit_totalEffect_binary, 
                     pars=c(HSCvar))[[1]]^2
model2_HSCvar_s2 <- 
  posterior_samples(fit_internalizedHomonegativity_binary, 
                     pars=c(HSCvar))[[1]]^2
model2_IHvar_s2 <- 
  posterior_samples(fit_internalizedHomonegativity_binary, 
                     pars=c(IHvar))[[1]]^2
model3_HSCvar_s2 <- 
  posterior_samples(fit_negativeExpectancies_binary, 
                     pars=c(HSCvar))[[1]]^2
model3_NEvar_s2 <- 
  posterior_samples(fit_negativeExpectancies_binary, 
                     pars=c(NEvar))[[1]]^2
model4_HSCvar_s2 <- 
  posterior_samples(fit_indirectOverall_binary, 
                     pars=c(HSCvar))[[1]]^2
model4_IHvar_s2 <- 
  posterior_samples(fit_indirectOverall_binary, 
                     pars=c(IHvar))[[1]]^2
model4_NEvar_s2 <- 
  posterior_samples(fit_indirectOverall_binary, 
                     pars=c(NEvar))[[1]]^2



# Create a matrix called "variance" to hold Level 2 variance estimates
variance_s2 <- matrix(nrow=8, ncol=1)
colnames(variance_s2) <- c("variance")
rownames(variance_s2) <-  c("Model 1: HSC",
                            "Model 2: HSC", 
                            "Model 2: IH",
                            "Modle 3: HSC",
                            "Model 3: NE",
                            "Model 4: HSC",
                            "Model 4: IH",
                            "Model 4: NE")
# assign variance values to each cell
variance_s2[1] <- round(mean(model1_HSCvar_s2),4)
variance_s2[2] <- round(mean(model2_HSCvar_s2),4)
variance_s2[3] <- round(mean(model2_IHvar_s2),4)
variance_s2[4] <- round(mean(model3_HSCvar_s2),4)
variance_s2[5] <- round(mean(model3_NEvar_s2),4)
variance_s2[6] <- round(mean(model4_HSCvar_s2),4)
variance_s2[7] <- round(mean(model4_IHvar_s2),4)
variance_s2[8] <- round(mean(model4_NEvar_s2),4)

```

## Calcluate indirect effect estimates and percentage mediated

```{r extractPosteriorDistributions_s2}

# Name of the primary parameter of interest
# Person mean-centered homonegative school climate
HSC <- "b_personMeanCenter_hscbinary"

# Extract estimates for this parameter from each model fit into numeric vector
model1_beta_samples_s2 <- 
  (posterior_samples(fit_totalEffect_binary, 
                    pars=c(HSC)))[[1]]
model2_beta_samples_s2 <- 
  posterior_samples(fit_internalizedHomonegativity_binary,
                    pars=c(HSC))[[1]]
model3_beta_samples_s2 <- 
  posterior_samples(fit_negativeExpectancies_binary, 
                    pars=c(HSC))[[1]]
model4_beta_samples_s2 <- 
  posterior_samples(fit_indirectOverall_binary, 
                    pars=c(HSC))[[1]]

# Use the difference method to calculate indirect effects
# subtract Model 2-4 beta post. distributions from Model 1 post. distributions
model2_difference_s2 <- model1_beta_samples_s2 - model2_beta_samples_s2
model3_difference_s2 <- model1_beta_samples_s2 - model3_beta_samples_s2
model4_difference_s2 <- model1_beta_samples_s2 - model4_beta_samples_s2

# calculate percentage mediated for each model (relative to Model 1)
model2_percent_s2 <- model2_difference_s2/model1_beta_samples_s2
model3_percent_s2 <- model3_difference_s2/model1_beta_samples_s2
model4_percent_s2 <- model4_difference_s2/model1_beta_samples_s2


# Create a matrix to hold indirect effect estimates
indirect_s2 <- as.data.frame(matrix(nrow=3, ncol=3))
colnames(indirect_s2) <- c("mean", "LB", "UB")
rownames(indirect_s2) <-  c("Model 2",
                            "Model 3",
                            "Model 4")
# Calculate indirect effect estimates for Models 2, 3, and 4
# Model 2
indirect_s2[1,1] <- exp(mean(model2_difference_s2))
indirect_s2[1,2:3] <- exp(quantile(model2_difference_s2, 
                                      probs=c(0.025,0.975)))
# Model 3
indirect_s2[2,1] <- exp(mean(model3_difference_s2))
indirect_s2[2,2:3] <- exp(quantile(model3_difference_s2, 
                                      probs=c(0.025,0.975)))
# Model 4
indirect_s2[3,1] <- exp(mean(model4_difference_s2))
indirect_s2[3,2:3] <- exp(quantile(model4_difference_s2, 
                                      probs=c(0.025,0.975)))
# combine means and 95% CI into one column
indirect_s2 <- indirect_s2 %>% 
  mutate(combined = 
           paste(
             format(round(mean,2),nsmall=2), 
             " (", 
             format(round(LB,2),nsmall=2),
             ", ", 
             format(round(UB,2),nsmall=2),
             ")", 
             sep=""))



# Create a matrix to hold percentage mediated estimates
percent_s2 <- as.data.frame(matrix(nrow=3, ncol=3))
colnames(percent_s2) <- c("mean", "LB", "UB")
rownames(percent_s2) <-  c("Model 2",
                           "Model 3",
                           "Model 4")
# Calculate percentage mediated for Models 2, 3, and 4
# Model 2
percent_s2[1,1] <- mean(model2_percent_s2)*100
percent_s2[1,2:3] <- quantile(model2_percent_s2, probs=c(0.025,0.975))*100
# Model 3
percent_s2[2,1] <- mean(model3_percent_s2)*100
percent_s2[2,2:3] <- quantile(model3_percent_s2, probs=c(0.025,0.975))*100
# Model 4 
percent_s2[3,1] <- mean(model4_percent_s2)*100
percent_s2[3,2:3] <- quantile(model4_percent_s2, probs=c(0.025,0.975))*100
# combine means and 95% CI into one column
percent_s2 <- percent_s2 %>% 
  mutate(combined = 
           paste(
             format(round(mean,1),nsmall=1), 
             "% (", 
             format(round(LB,1),nsmall=1),
             ", ", 
             format(round(UB,1),nsmall=1),
             ")", 
             sep=""))

```

## Create and Export Table S2

```{r tableS2}

# parameter names
tableS2_params <-   c("Model 1",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "",
                      "Model 2",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "PMC Internalized homonegativity, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)",
                      "",
                      "Model 3",
                      "PMC Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "PMC Negative expectancies, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)",
                      "",
                      "Model 4",
                      "Homonegative school climate, RR (95% CI)",
                      "    Random slopes variance",
                      "Internalized homonegativity, RR (95% CI)",
                      "    Random slopes variance",
                      "Negative expectancies, RR (95% CI)",
                      "    Random slopes variance",
                      "Percentage of total effect mediated (95% CI)")
# model estimates
tableS2_estimates <-  c("",
                        fixedEffects_s2[1,4],
                        variance_s2[1],
                        "",
                        "",
                        fixedEffects_s2[3,4],
                        variance_s2[2],
                        fixedEffects_s2[4,4],
                        variance_s2[3],
                        percent_s2[1,4],
                        "",
                        "",
                        fixedEffects_s2[7,4],
                        variance_s2[4],
                        fixedEffects_s2[8,4],
                        variance_s2[5],
                        percent_s2[2,4],
                        "",
                        "",
                        fixedEffects_s2[11,4],
                        variance_s2[6],
                        fixedEffects_s2[12,4],
                        variance_s2[7],
                        fixedEffects_s2[13,4],
                        variance_s2[8],
                        percent_s2[3,4])
# combine columns into one table
tableS2 <- as.data.frame(cbind(tableS2_params, tableS2_estimates))
# change column names
colnames(tableS2) <- c("Parameters", "Estimates")

# load library
library(openxlsx)

# save xlsx file
openxlsx::write.xlsx(tableS2, file="supplemental_files/tableS2.xlsx")
# reload into R in xlsx format
wb <- openxlsx::loadWorkbook(file="supplemental_files/tableS2.xlsx")
# set column width for PARAMETER to 40
openxlsx::setColWidths(wb, 
                       sheet = 1,
                       cols = 1, 
                       widths=50)
# set column width for model estimates to 20
openxlsx::setColWidths(wb, 
                       sheet = 1,
                       cols = 2, 
                       widths=30)
# change font size to 10 with Arial style
modifyBaseFont(wb, 
               fontSize = 10, 
               fontColour = "black", 
               fontName = "Times New Roman")

# bold text
boldStyle <- createStyle(textDecoration = "Bold")
# center-aligned text
centerStyle <- createStyle(halign = "CENTER")
# bold and center-aligned
boldCenterStyle <- createStyle(textDecoration = "Bold",
                               halign = "CENTER")

# apply styles
addStyle(wb,
         rows = c(1,2,6,13,20),
         cols = 1,
         sheet = 1,
         style = boldStyle)
addStyle(wb,
         cols = 2,
         rows = 2:27,
         sheet = 1,
         style = centerStyle,
         gridExpand = TRUE)
addStyle(wb,
         rows = c(1,3,7,9,14,16,21,23,25),
         cols = 2,
         sheet = 1,
         style = boldCenterStyle,
         gridExpand = TRUE)

# save the final excel workbook
openxlsx::saveWorkbook(wb, 
                         file="supplemental_files/tableS2.xlsx", 
                       overwrite = TRUE)

```

# Respecify "probable BDD" variable to include thin/fat concerns

This code calculates the point and study period prevalence estimates of
the "probable BDD" variable specification which includes those who
report that their main concern was that they are too fat or not thin
enough.

In the main analysis, these participants were coded into the reference
category.

```{r}

# load data
imputed_long_cisgender <- readRDS("imputed_long_cisgender.rds")

# create matrix to store values
probableBDD_thinfat <- as.data.frame(matrix(nrow=4, ncol=4))

# Probable BDD
# time 0 (baseline)
bdd_0 <- summary(pool(with(imputed_long_cisgender, 
                           lm(bddbinThinFat*100~1, subset=time==0))))
probableBDD_thinfat[1,1] <- bdd_0[2] # mean
probableBDD_thinfat[1,2] <- bdd_0[2]-1.96*bdd_0[3] # mean - 1.96*SE
probableBDD_thinfat[1,3] <- bdd_0[2]+1.96*bdd_0[3] # mean + 1.96*SE
# time 1 (6 month)
bdd_1 <- summary(pool(with(imputed_long_cisgender, 
                           lm(bddbinThinFat*100~1, subset=time==0.5))))
probableBDD_thinfat[2,1] <- bdd_1[2] # mean
probableBDD_thinfat[2,2] <- bdd_1[2]-1.96*bdd_1[3] # mean - 1.96*SE
probableBDD_thinfat[2,3] <- bdd_1[2]+1.96*bdd_1[3] # mean + 1.96*SE
# time 2 (12 month)
bdd_2 <- summary(pool(with(imputed_long_cisgender, 
                           lm(bddbinThinFat*100~1, subset=time==1))))
probableBDD_thinfat[3,1] <- bdd_2[2] # mean
probableBDD_thinfat[3,2] <- bdd_2[2]-1.96*bdd_2[3] # mean - 1.96*SE
probableBDD_thinfat[3,3] <- bdd_2[2]+1.96*bdd_2[3] # mean + 1.96*SE
# overall (at any time point)
bdd_3 <- summary(pool(with(imputed_long_cisgender, 
                           lm(bddbinThinFat_overall*100~1))))
probableBDD_thinfat[4,1] <- bdd_3[2] # mean
probableBDD_thinfat[4,2] <- bdd_3[2]-1.96*bdd_3[3] # mean - 1.96*SE
probableBDD_thinfat[4,3] <- bdd_3[2]+1.96*bdd_3[3] # mean + 1.96*SE


# round to 2 decimal places
probableBDD_thinfat <- round(probableBDD_thinfat,2)
# create cell with mean + 95% CI
combine <- function(row) {
  probableBDD_thinfat[row,4] <- paste(probableBDD_thinfat[row,1]," (",
                                      probableBDD_thinfat[row,2],", ",
                                      probableBDD_thinfat[row,3],")", sep="")
}

# run the above function to combine the estimates into one cell
probableBDD_thinfat[1,4] <- combine(row=1)
probableBDD_thinfat[2,4] <- combine(row=2)
probableBDD_thinfat[3,4] <- combine(row=3)
probableBDD_thinfat[4,4] <- combine(row=4)

# rename columns and rows
col1 <- c("Baseline point prevalence",
          "6-month point prevalence",
          "12-month point prevalence",
          "Study period prevalence")
probableBDD_thinfat <- probableBDD_thinfat %>% 
  mutate(col1 = col1, .before=1)
# rename columns
colnames(probableBDD_thinfat) <- c("Time point",
                                   "Mean",
                                   "Lower bound",
                                   "Upper bound",
                                   "Mean (95% CI)")

# export the table to excel
library(openxlsx)
openxlsx::write.xlsx(probableBDD_thinfat, file="supplemental_files/probableBDD_thinfat.xlsx")
```
